Refactored SPI integration tests to use RepositoryContainerBuilder
